# app/frontend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/verification/claims/{id}": {
      "get": {
        "description": "Retrieve the current verification status and details of a claim",
        "operationId": "VerificationController_findClaim_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Unique identifier of the claim",
            "schema": {
              "example": "clv789xyz123",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Claim verification status retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "clv789xyz123",
                    "status": "verified",
                    "verificationScore": 0.85,
                    "verificationResult": {
                      "score": 0.85,
                      "confidence": 0.92,
                      "details": {
                        "factors": [
                          "Document authenticity verified",
                          "Identity cross-reference passed"
                        ],
                        "riskLevel": "low"
                      },
                      "processedAt": "2025-01-23T14:30:00.000Z"
                    },
                    "verifiedAt": "2025-01-23T14:30:00.000Z"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - valid JWT token required."
          },
          "404": {
            "description": "The specified claim could not be found."
          }
        },
        "security": [
          {
            "x-api-key": []
          },
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get claim verification status",
        "tags": [
          "Verification"
        ]
      }
    },
...
    "/api/v1/claims": {
      "post": {
        "description": "Initializes a new claim for a specific campaign.",
        "operationId": "ClaimsController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClaimDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Claim created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateClaimDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input parameters."
          },
          "404": {
            "description": "The specified campaign was not found."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a claim",
        "tags": [
          "Onchain Proxy"
        ]
      },
      "get": {
        "description": "Retrieves a list of all claims across all campaigns.",
        "operationId": "ClaimsController_findAll_v1",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items to return (default: 25, max: 100)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque cursor for pagination",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResult"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/CreateClaimDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "List all claims",
        "tags": [
          "Onchain Proxy"
        ]
      }
    },
    "/api/v1/claims/{id}": {
      "get": {
        "description": "Retrieves the current details and status of a specific claim.",
        "operationId": "ClaimsController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Claim details retrieved successfully."
          },
          "404": {
            "description": "The specified claim was not found."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get claim details",
        "tags": [
          "Onchain Proxy"
        ]
      }
    },
    "/api/v1/claims/{id}/verify": {
      "post": {
        "description": "Marks a claim as verified. Requires operator or admin role.",
        "operationId": "ClaimsController_verify_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Claim status transitioned to verified successfully."
          },
          "400": {
            "description": "Invalid status transition."
          },
          "403": {
            "description": "Access denied - insufficient permissions."
          },
          "404": {
            "description": "The specified claim was not found."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Verify a claim",
        "tags": [
          "Onchain Proxy"
        ]
      }
    },
    "/api/v1/claims/{id}/approve": {
      "post": {
        "description": "Approves a verified claim. Requires admin role.",
        "operationId": "ClaimsController_approve_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Claim approved successfully (verified → approved)."
          },
          "400": {
            "description": "Invalid status transition."
          },
          "403": {
            "description": "Access denied - admin role required."
          },
          "404": {
            "description": "The specified claim was not found."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Approve a claim",
        "tags": [
          "Onchain Proxy"
        ]
      }
    },
    "/api/v1/claims/{id}/disburse": {
      "post": {
        "description": "Initiates on-chain disbursement for an approved claim. Requires admin role.",
        "operationId": "ClaimsController_disburse_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "On-chain disbursement initiated or completed successfully.",
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "summary": "Successful on-chain disbursement",
                    "value": {
                      "id": "claim_123",
                      "status": "disbursed",
                      "transactionHash": "0x123...abc",
                      "amount": "100.50"
                    }
                  },
                  "pending": {
                    "summary": "Disbursement pending on-chain",
                    "value": {
                      "id": "claim_123",
                      "status": "disbursing",
                      "message": "Check back for final transaction hash."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid status transition or account state."
          },
          "403": {
            "description": "Access denied - admin role required."
          },
          "404": {
            "description": "The specified claim was not found."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Disburse funds for a claim",
        "tags": [
          "Onchain Proxy"
        ]
      }
    },
    "/api/v1/claims/{id}/archive": {
      "patch": {
        "description": "Soft-archives a claim, hiding it from general listings.",
        "operationId": "ClaimsController_archive_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Claim archived successfully."
          },
          "400": {
            "description": "Invalid status transition."
          },
          "404": {
            "description": "The specified claim was not found."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Archive a claim",
        "tags": [
          "Onchain Proxy"
        ]
      }
    },
    "/api/v1/claims/{id}/notes": {
      "post": {
        "description": "Adds a secure internal note for staff review only.",
        "operationId": "ClaimsController_addNote_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInternalNoteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Internal note added successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalNoteResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Access denied - staff role required."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Add an internal note to a claim",
        "tags": [
          "Onchain Proxy"
        ]
      },
      "get": {
        "description": "Retrieves all internal notes for a specific claim.",
        "operationId": "ClaimsController_getNotes_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Internal notes retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InternalNoteResponseDto"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Access denied - staff role required."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "List internal notes for a claim",
        "tags": [
          "Onchain Proxy"
        ]
      }
    },
    "/api/v1/claims/{id}/cancel": {
      "post": {
        "description": "Cancels an active claim (requested / verified / approved). Releases the locked budget back to the campaign and records a full audit trail. Disbursed claims cannot be cancelled.",
        "operationId": "ClaimsController_cancel_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelClaimDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Claim cancelled successfully."
          },
          "400": {
            "description": "Claim is already cancelled or in a non-cancellable status."
          },
          "403": {
            "description": "Access denied - operator role required."
          },
          "404": {
            "description": "Claim not found."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Cancel a claim",
        "tags": [
          "Onchain Proxy"
        ]
      }
    },
    "/api/v1/claims/{id}/reissue": {
      "post": {
        "description": "Atomically cancels the original claim and creates a replacement. The replacement is linked to the original via `reissuedFromId`, preserving the full audit chain. Locked balances are transferred to the new claim — no double-counting occurs. Returns both the cancelled original and the new replacement.",
        "operationId": "ClaimsController_reissue_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReissueClaimDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Original claim cancelled and replacement created.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "original": {
                      "type": "object",
                      "description": "The cancelled original claim."
                    },
                    "replacement": {
                      "type": "object",
                      "description": "The newly created replacement claim."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Original claim is not in a cancellable status."
          },
          "403": {
            "description": "Access denied - operator role required."
          },
          "404": {
            "description": "Original claim not found."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Cancel and reissue a claim",
        "tags": [
          "Onchain Proxy"
        ]
      }
    },
    "/api/v1/claims/{id}/reissue-history": {
      "get": {
        "description": "Returns the full lineage of a claim — the original and every replacement — ordered from oldest to newest. Pass any claim ID in the chain to retrieve the complete history.",
        "operationId": "ClaimsController_getReissueHistory_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reissue chain retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Access denied - operator role required."
          },
          "404": {
            "description": "Claim not found."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get reissue chain for a claim",
        "tags": [
          "Onchain Proxy"
        ]
      }
    },
...
    "/api/v1/entity-linking/claim/{claimId}": {
      "get": {
        "description": "Retrieve all entity links associated with a specific claim",
        "operationId": "EntityLinkingController_getLinksByClaim_v1",
        "parameters": [
          {
            "name": "claimId",
            "required": true,
            "in": "path",
            "description": "Claim ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entityType",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "organization",
                "location",
                "asset",
                "project"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get entity links by claim",
        "tags": [
          "Entity Linking"
        ]
      }
    },
...
  },
  "info": {
    "title": "ChainForge API",
    "description": "API documentation for ChainForge platform - Humanitarian aid and verification system\n\n## API Versioning\n\nThis API uses URI-based versioning. The current version is **v1**.\n\n### Version Format\nAll endpoints are prefixed with the version number: `/api/v1/...`\n\n### Supported Versions\n| Version | Status | Description |\n|---------|--------|-------------|\n| v1 | Current | Active version with full support |\n\n### Deprecation Policy\n- Deprecated endpoints will be marked with `@Deprecated` in the documentation\n- Deprecated versions will be supported for at least 6 months after deprecation notice\n- Clients will receive deprecation warnings via the `Sunset` HTTP header\n- Migration guides will be provided for major version changes\n\n### Future Versions\nWhen new versions are released:\n- New endpoints will be available at `/api/v2/...`, etc.\n- Previous versions remain accessible during the deprecation period\n- Clients should monitor the API documentation for version updates",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development"
    },
    {
      "url": "https://api.chainforge.app",
      "description": "Staging"
    },
    {
      "url": "https://api.chainforge.app",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "in": "header",
        "description": "Enter JWT token"
      },
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key for external access"
      }
    },
    "schemas": {
...
      "CreateInternalNoteDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "description": "The content of the internal note.",
            "example": "Investigation shows that the recipient has provided valid evidence."
          },
          "category": {
            "type": "string",
            "description": "Optional category for the note.",
            "example": "investigation"
          }
        },
        "required": [
          "content"
        ]
      },
      "InternalNoteResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "clv789xyz123"
          },
          "entityType": {
            "type": "string",
            "example": "claim"
          },
          "entityId": {
            "type": "string",
            "example": "clv789xyz123"
          },
          "content": {
            "type": "string",
            "example": "Investigation notes..."
          },
          "authorId": {
            "type": "string",
            "example": "clv789xyz123"
          },
          "category": {
            "type": "object",
            "example": "investigation",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-01-23T11:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-01-23T11:00:00.000Z"
          }
        },
        "required": [
          "id",
          "entityType",
          "entityId",
          "content",
          "authorId",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "PaginatedResult": {
        "type": "object",
        "properties": {
          "data": {
            "description": "The array of records for the current page.",
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true,
            "description": "Cursor to retrieve the next page of results, or null if there is no next page."
          }
        },
        "required": [
          "data"
        ]
      },
...
      "CreateClaimDto": {
        "type": "object",
        "properties": {
          "campaignId": {
            "type": "string",
            "description": "ID of the campaign this claim belongs to",
            "example": "campaign-uuid"
          },
          "amount": {
            "type": "number",
            "description": "Amount requested in the claim",
            "example": 100.5,
            "minimum": 0
          },
          "recipientRef": {
            "type": "string",
            "description": "Reference to the recipient",
            "example": "recipient-123"
          },
          "tokenAddress": {
            "type": "string",
            "description": "Stellar token address (asset issuer or contract ID) for the distribution",
            "example": "GATEMHCCKCY67ZUCKTROYN24ZYT5GK4EQZ5LKG3FZTSZ3NYNEJBBENSN"
          },
          "evidenceRef": {
            "type": "string",
            "description": "Reference or link to evidence supporting the claim (e.g., photo, document hash).",
            "example": "evidence-456"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string",
            "description": "When the claim should automatically expire if it remains unprocessed.",
            "example": "2026-05-31T23:59:59.000Z"
          }
        },
        "required": [
          "campaignId",
          "amount",
          "recipientRef",
          "tokenAddress"
        ]
      },
...
    }
  }
}